news_summariser/requirements.txt
fastapi==0.115.12
uvicorn==0.20.0
pydantic==2.11.2
pydantic_core==2.33.1
python-dotenv==1.0.1
requests==2.31.0
newspaper3k==0.2.8
nltk==3.8.1 